requirements.txt
tiktoken>=0.5.1
gitpython>=3.1.40
jinja2>=3.1.2
radon>=6.0.1
ruff>=0.1.9
requirements-parser>=0.5.0
packaging>=23.2
